https://editor.p5js.org/ss14740/sketches/El5Z2IuoL
For this assignment, I enjoyed playing around with different shapes and creating loops. I was able to get a clearer understanding of what the variables i and j mean and as well as how to manipulate rectangles and lines into the position and size I want them in. I struggled with incorporating more than one shape, as I tried so add a circle but was unable to due to repeated errors.